+++ src/engine/player_response.rs
@@ -8,6 +8,6 @@ pub struct PlayerResponse {
     pub player: Player,
     pub piece: Piece,
     pub raw_response: Option<String>,
-    pub placement_count: u32,
+    pub placement_count: usize,
     pub error: Option<String>,
 }
